+++ AWSKMS/AWSKMSService.m
@@ -25,7 +25,7 @@
 #import "AWSKMSResources.h"
 
 static NSString *const AWSInfoKMS = @"KMS";
-NSString *const AWSKMSSDKVersion = @"2.11.1";
+NSString *const AWSKMSSDKVersion = @"2.12.0";
 
 
 @interface AWSKMSResponseSerializer : AWSJSONResponseSerializer